# server/api/openapi.yaml
openapi: 3.0.3
info:
  title: AgentTeams API
  description: Enterprise Windows Machine Management System API
  version: 1.0.0
  contact:
    name: AgentTeams Team
servers:
  - url: https://api.agentteams.io/api/v1
    description: Production server
...
paths:
...
  /agents:
    get:
      tags: [Agents]
      summary: List all agents
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [online, offline, busy]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
    post:
      tags: [Agents]
      summary: Register new agent
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
  /agents/{id}:
    get:
      tags: [Agents]
      summary: Get agent by ID
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentResponse'
        '404':
          description: Agent not found
    delete:
      tags: [Agents]
      summary: Delete agent
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Agent deleted
...
  /versions:
    get:
      tags: [Versions]
      summary: List versions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of versions
    post:
      tags: [Versions]
      summary: Upload new version
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/UploadVersionRequest'
      responses:
        '201':
          description: Version uploaded
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateAgentRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          maxLength: 100
        metadata:
          type: object
    AgentResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        status:
          type: string
          enum: [online, offline, busy]
        version:
          type: string
        last_seen_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    AgentListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AgentResponse'
        pagination:
          $ref: '#/components/schemas/Pagination'
...
    UploadVersionRequest:
      type: object
      properties:
        version:
          type: string
        file:
          type: string
          format: binary
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer